requirements.txt
djangorestframework
corsheaders
rest_framework_simplejwt
drf_social_oauth2


